Lab4 - Implemented users rating

+++ app/src/main/java/com/example/group15_lab2/DataClasses/FilterParams.kt
@@ -2,6 +2,6 @@ package com.example.group15_lab2.DataClasses
 
 data class FilterParams (
     var category:String = "ALL",
-    var from_price:Int? = null,
-    var to_price:Int? = null
+    var from_price:Float? = null,
+    var to_price:Float? = null
 )
